pciemgr: OCP portmap overflow (#27999)
The portmap register has 48 active bits, but pciemgr wrote with
32-bit data types truncating the port bits for upper lifs 74-79.
OCP cards use port4 so need a 4 set in the portmap register.
All other cards use port0 so default of 0 was correct.

+++ nic/sdk/platform/pciemgr/src/portmap.c
@@ -54,8 +54,8 @@ portmap_setlif(const u_int32_t lif, const u_int8_t port)
     u_int64_t e;
 
     portmap_get(idx, &e);
-    e &= ~(0x7 << (off * 3));
-    e |= (port & 0x7) << (off * 3);
+    e &= ~((u_int64_t)0x7 << (off * 3));
+    e |= ((u_int64_t)port & 0x7) << (off * 3);
     portmap_set(idx, e);
 }
